feat: [#19] MainItemView xib 생성
