--- FullStackCodingBot/FullStackCodingBot/View/CustomView/CustomView.swift
@@ -1,23 +1,0 @@
-import UIKit
-
-class CustomView: UIView {
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        loadXib()
-    }
-    
-    required init?(coder: NSCoder) {
-        super.init(coder: coder)
-        loadXib()
-    }
-    
-    private func loadXib() {
-        let identifier = String(describing: type(of: self))
-        let nibs = Bundle.main.loadNibNamed(identifier, owner: self, options: nil)
-        guard let customView = nibs?.first as? UIView else { return }
-        customView.frame = self.bounds
-        self.addSubview(customView)
-    }
-}
-
